server/requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.29
langchain_experimental==0.0.64
langchain-huggingface==0.0.3
langchain-postgres==0.0.9
langchain-text-splitters==0.2.1
langchain-google-genai==1.0.7
langchain-openai==0.1.9
langchain-ollama==0.1.1
sentence-transformers==2.6.1
transformers==4.43.1
bitsandbytes==0.42.0
torch==2.3.1
fastapi
uvicorn
markdown
python-dotenv==1.0.1
docx2txt==0.8
unstructured==0.14.3
unstructured-client==0.22.0
XlsxWriter==3.2.0
python-pptx==0.6.23
openpyxl==3.1.3
tqdm
lxml==5.1.0
python-dotenv
rank_bm25
openai==1.35.3
google-ai-generativelanguage==0.6.6
pillow==10.4.0
ragas==0.1.15
jq
flashrank
accelerate==0.34.0
pydantic==2.7.4
pydantic_core==2.18.4
psycopg2-binary==2.9.9